repoman: Move thirdpartymirror check to the fetches module
This check should be part of the FetchChecks class.
This removes 'src_uri_error' Future need.
